# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Auto Dealer Operator API
  version: 4.0.0
  description: |
    Contract for the operator publish console and dealer inventory APIs.
    Handlers are implemented in Fastify; this spec is the source for the generated web client.
servers:
  - url: /
    description: Same-origin (Vite proxy in dev)
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      x-route-classification: auth
      operationId: operatorLogin
      summary: Operator login — issues op_session HttpOnly cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorIdentityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/me:
    get:
      tags: [Auth]
      security:
        - CookieAuth: []
      x-route-classification: auth
      operationId: getOperatorMe
      summary: Get current operator identity
      responses:
        '200':
          description: Authenticated operator identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorIdentityResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/dealers:
    get:
      tags: [Dealers]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: listDealers
      summary: List dealerships
      responses:
        '200':
          description: Dealer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Dealers]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: createOperatorDealership
      summary: Create a dealership profile and grant the current operator access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealershipRequest'
      responses:
        '201':
          description: Created dealership profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDealershipResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate dealership
  /api/dealers/signup:
    post:
      tags: [Dealers]
      security: []
      x-route-classification: public-write
      operationId: createDealershipSignup
      summary: Create a dealership profile from the public signup flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealershipRequest'
      responses:
        '201':
          description: Created dealership profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDealershipResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Duplicate dealership
  /api/dealers/signup/{dealershipId}/logo:
    post:
      tags: [Dealers]
      security: []
      x-route-classification: public-write
      operationId: uploadSignupDealerLogo
      summary: Upload logo image during public dealership signup
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [logo]
              properties:
                logo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Uploaded logo URL
          content:
            application/json:
              schema:
                type: object
                required: [logoUrl]
                properties:
                  logoUrl:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/publish/prepare:
    post:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: preparePublish
      summary: Prepare publish (dry-run by default)
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreparePublishRequest'
      responses:
        '200':
          description: Prepare result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreparePublishResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/dealers/{dealershipId}/publish/status:
    get:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getPublishStatus
      summary: Publish readiness status (dry-run prepare)
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Status snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/publish/history:
    get:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getPublishHistory
      summary: Sync event history
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - name: platformSlug
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: before
          in: query
          description: Cursor sync event id
          schema:
            type: string
      responses:
        '200':
          description: History page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/publish/accounts:
    get:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getPublishAccounts
      summary: Platform account summary for publish console
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Account rows aligned to platform catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishAccountsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/publish/queue:
    get:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getPublishQueue
      summary: Publish queue view
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Queue buckets and summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueView'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/publish/auto-sync:
    get:
      tags: [Publish]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getAutoSyncStatus
      summary: Auto-sync daemon state for a dealer
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Current auto-sync phase and last run info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoSyncStatus'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/accounts:
    get:
      tags: [Accounts]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: listAccounts
      summary: Platform accounts (management)
      parameters:
        - $ref: '#/components/parameters/dealershipId'
      responses:
        '200':
          description: Accounts with summary counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsManagementResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/accounts/{platformSlug}:
    patch:
      tags: [Accounts]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: updateAccount
      summary: Update platform account fields
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdatePayload'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/platforms/{platformSlug}/connect-url:
    get:
      tags: [Accounts]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getPlatformConnectUrl
      summary: Get OAuth authorization URL for a platform
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      responses:
        '200':
          description: Authorization URL and CSRF state token
          content:
            application/json:
              schema:
                type: object
                required: [authUrl, state]
                properties:
                  authUrl:
                    type: string
                    description: Redirect the user to this URL to begin OAuth authorization
                  state:
                    type: string
                    description: CSRF state token stored server-side
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/platforms/{platformSlug}/oauth-token:
    delete:
      tags: [Accounts]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: deletePlatformOAuthToken
      summary: Revoke stored OAuth token for a platform
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      responses:
        '204':
          description: Token deleted
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/dealers/{dealershipId}/platforms/{platformSlug}/pages:
    get:
      tags: [Social]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getSocialPages
      summary: Sync and list connected social pages for a platform
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      responses:
        '200':
          description: List of connected pages (no access tokens exposed)
          content:
            application/json:
              schema:
                type: object
                required: [pages]
                properties:
                  pages:
                    type: array
                    items:
                      $ref: '#/components/schemas/SocialPageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: OAuth not connected — reconnect first
        '502':
          description: Upstream Graph API error
  /api/dealers/{dealershipId}/platforms/{platformSlug}/pages/{pageId}/select:
    put:
      tags: [Social]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: selectSocialPage
      summary: Select a social page for this dealer
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated page list with new selection reflected
          content:
            application/json:
              schema:
                type: object
                required: [pages]
                properties:
                  pages:
                    type: array
                    items:
                      $ref: '#/components/schemas/SocialPageSummary'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/platforms/{platformSlug}/posts/preview:
    post:
      tags: [Social]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: previewSocialPost
      summary: Preview generated post copy for a vehicle (no publish, no DB write)
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId]
              properties:
                vehicleId:
                  type: string
      responses:
        '200':
          description: Preview of post text, image, listing URL, and selected page
          content:
            application/json:
              schema:
                type: object
                required: [preview]
                properties:
                  preview:
                    $ref: '#/components/schemas/PostPreview'
                  selectedPage:
                    nullable: true
                    $ref: '#/components/schemas/SocialPageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/dealers/{dealershipId}/platforms/{platformSlug}/posts:
    post:
      tags: [Social]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: publishSocialPost
      summary: Publish a vehicle post to the selected social page
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId]
              properties:
                vehicleId:
                  type: string
                trigger:
                  type: string
                  enum: [MANUAL, AUTO]
                  default: MANUAL
                source:
                  type: string
      responses:
        '201':
          description: Post published and stored
          content:
            application/json:
              schema:
                type: object
                required: [post]
                properties:
                  post:
                    $ref: '#/components/schemas/SocialPost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No page selected — choose a page first
        '502':
          description: Upstream Graph API publish error
    get:
      tags: [Social]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: listSocialPosts
      summary: List recent social posts for a dealer platform
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
      responses:
        '200':
          description: List of recent posts, newest first
          content:
            application/json:
              schema:
                type: object
                required: [posts]
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SocialPost'
...
  /api/dealers/{dealershipId}/platforms/{platformSlug}/listings/{vehicleId}:
    get:
      tags: [Listings]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: getMarketplaceListing
      summary: Get the listing record for one vehicle on a platform
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
        - $ref: '#/components/parameters/vehicleId'
      responses:
        '200':
          description: Listing record
          content:
            application/json:
              schema:
                type: object
                required: [listing]
                properties:
                  listing:
                    $ref: '#/components/schemas/MarketplaceListing'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Listings]
      security:
        - OperatorAuth: []
      x-route-classification: operator
      operationId: endMarketplaceListing
      summary: Withdraw the offer and mark the listing ENDED
      parameters:
        - $ref: '#/components/parameters/dealershipId'
        - $ref: '#/components/parameters/platformSlug'
        - $ref: '#/components/parameters/vehicleId'
      responses:
        '204':
          description: Listing ended
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Platform OAuth not connected — reconnect first
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: No external offer ID — nothing to withdraw
        '502':
          description: Platform API error withdrawing the offer
  /api/oauth/callback:
    get:
      tags: [Accounts]
      security: []
      x-route-classification: public
      operationId: oauthCallback
      summary: OAuth provider callback — exchanges authorization code for tokens
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: error
          in: query
          schema:
            type: string
        - name: error_description
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTML popup-close page (success or error)
          content:
            text/html:
              schema:
                type: string
        '400':
          description: HTML popup-close page (invalid state)
          content:
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    OperatorAuth:
      type: apiKey
      in: cookie
      name: op_session
      description: |
        Operator session cookie issued by POST /api/auth/login (8-hour lifetime, HttpOnly, Secure in production).
        In development/test the Fastify guard also accepts an x-operator-id request header as a fallback;
        that path is never active when NODE_ENV=production.
    CookieAuth:
      type: apiKey
      in: cookie
      name: op_session
      description: HttpOnly session cookie issued by POST /api/auth/login. Secure flag set in production.
  parameters:
    dealershipId:
      name: dealershipId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    OperatorLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
    OperatorIdentityResponse:
      type: object
      required: [id, email, role, dealerAccessIds]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, OPERATOR, DEALER_OPERATOR]
        dealerAccessIds:
          type: array
          items:
            type: string
          description: Dealership IDs the operator can access. Empty for SUPER_ADMIN (global access).
...